resolved code smells
