+++ dimmer-core/src/main/java/com/github/cloudyrock/dimmer/ValueFeatureMetadata.java
@@ -1,7 +1,5 @@
 package com.github.cloudyrock.dimmer;
 
-import java.util.Objects;
-
 final class ValueFeatureMetadata extends FeatureMetadata {
 
     private final Object valueToReturn;
